migrations/pyproject.toml
[project]
name = "pgwarden"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg[binary]>=3.3.3",
    "yoyo-migrations>=9.0.0",
]
